nice refactoring in backend  and add cloud deleting status

+++ backend/users/src/main/java/com/wordium/users/dto/Stats.java
@@ -1,7 +1,7 @@
 package com.wordium.users.dto;
       public record Stats(
-                        int followers,
-                        int following,
-                        int posts,
-                        int bookmarks) {
+                        Long followers,
+                        Long following,
+                        Long posts,
+                        Long bookmarks) {
         }
